Add files via upload
The random walk algorithm is utilized to teach concepts of randomness, probability, and statistical modeling. Students can grasp the idea of unpredictable movement and how patterns emerge from randomness. It serves as a practical example of mathematical concepts in real-world scenarios.

In summary, the random walk algorithm is a versatile concept used across various disciplines to model stochastic processes. It finds applications in physics, biology, computer science, and more. Its educational value lies in its ability to teach probability, randomness, and simulation techniques to students, making it a valuable tool for understanding complex systems and processes.